+++ MOP/src/Rules/Types/SpecialRules.cs
@@ -25,7 +25,6 @@ namespace MOP.Rules.Types
         #region Custom Rule Files
         public bool IgnoreModVehicles;
         public bool ToggleAllVehiclesPhysicsOnly;
-        public bool ShowGarbageMemoryUsage;
         #endregion
     }
 }