+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.8.3 (TBA)
 
+### Added
+
+- Added command "mop debug_monitor", which shows debug stats of MOP
+
 ### Bug Fixes
 
 - MOP won't show errors during save file verification, if the new save file has been created, but  the game hasn't been saved, which made certain save tags missing
@@ -10,6 +14,7 @@
 ### Removed
 
 - Removed unused code
+- Removed custom rule file flag: "show_garbage_memory_usage"
 
 ## 3.8.2 (18.08.2022)